Update framework, packages, and improve documentation
Updated the target framework to .NET 8.0 and added `SdkAnalysisLevel` in the project files.
Updated several NuGet package references to newer versions in `EF_StringEncryptPropertyValues.csproj` and `MiscSettingsApp.csproj`.
Removed unused `using` directives in `Operations.cs` and `Program.cs`.
Improved XML documentation comments in `JSonHelper.cs` for better clarity.
Refactored JSON serialization methods in `JSonHelper.cs` to use a common `SerializerOptions` property.
Corrected an enum value in `appsettings.json` from "One1" to "One".
Updated the `readme.md` file to include additional explanations and notes about the project, including a note on the importance of validation and a quirk with reading enum values.
Added a new section in `readme.md` to explain the validation setup and the expected failure due to an invalid enum value.
Updated image references and added a note about the `ValidateOnStart` method in the `readme.md`.
Removed redundant package references and updated to the latest versions where applicable.